examples/images/vit/requirements.txt
thop
torchprofile
torchsummary
regex
wandb
huggingface_hub
tokenizers
fvcore
